Add files via upload
After finding the universal variable m_size I was able to successfully compete the size function.

For the search function I ran into a road block when trying to compare node temp with value. After looking through the rest of the code and finding the getValue function I was able to successfully complete this function.

The removeBack function took me the most time to complete. At first I wasn't sure why we needed 2 node variable to use. After reviewing my notes I remembered that we needed the second variable to keep the original m_front was correct. Once this was realized the rest of the function worked smoothly and was completed.

Overall the lab took about 3 hours to complete and was a great review for getting back into c++ coding.